feat(entity-id-type): modifica o tipo da id para Big Integer

+++ docs/db/schema.sql
@@ -1,11 +1,11 @@
 CREATE TABLE departamento (
-	id_departamento INTEGER AUTO_INCREMENT,
+	id_departamento BIGINT AUTO_INCREMENT,
 	nm_departamento VARCHAR(50) NOT NULL,
 	PRIMARY KEY(id_departamento)
 );
 
 CREATE TABLE funcionario (
-	id_funcionario INTEGER AUTO_INCREMENT,
+	id_funcionario BIGINT AUTO_INCREMENT,
 	nome VARCHAR(100) NOT NULL,
 	cpf CHAR(11) NOT NULL UNIQUE,
 	email VARCHAR(100) NOT NULL UNIQUE,
@@ -19,8 +19,8 @@ CREATE TABLE funcionario (
 	logradouro VARCHAR(100) NOT NULL,
 	numero VARCHAR(10),
 	cep CHAR(9),
-	id_departamento INTEGER,
-	criado_por INTEGER,
+	id_departamento BIGINT,
+	criado_por BIGINT,
 	PRIMARY KEY(id_funcionario),
 	FOREIGN KEY (id_departamento) REFERENCES departamento(id_departamento),
 	FOREIGN KEY (criado_por) REFERENCES funcionario(id_funcionario)
