+++ src/main/java/com/delegrego/exemplo_spring_boot_2/controller/FuncionarioController.java
@@ -1,5 +1,6 @@
 package com.delegrego.exemplo_spring_boot_2.controller;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -73,7 +74,7 @@ public class FuncionarioController {
 	 * @return ResponseEntity com o funcionário encontrado
 	 */
 	@GetMapping("/{id}")
-	public ResponseEntity<FuncionarioResponseDto> obterFuncionarioPorId(@PathVariable int id) {
+	public ResponseEntity<FuncionarioResponseDto> obterFuncionarioPorId(@PathVariable BigInteger id) {
 		return ResponseEntity.status(HttpStatus.OK).body(servico.obterFuncionarioPorId(id));
 	}
 
@@ -96,7 +97,7 @@ public class FuncionarioController {
 	 * @return ResponseEntity com status apropriado
 	 */
 	@PutMapping("/{id}")
-	public ResponseEntity<Void> atualizarFuncionario(@PathVariable int id,
+	public ResponseEntity<Void> atualizarFuncionario(@PathVariable BigInteger id,
 			@Valid @RequestBody FuncionarioAtualizarDto funcionarioDto) {
 
 		servico.atualizarFuncionario(id, funcionarioDto);
@@ -111,7 +112,7 @@ public class FuncionarioController {
 	 * @return ResponseEntity com status apropriado
 	 */
 	@DeleteMapping("/{id}")
-	public ResponseEntity<Void> deletarFuncionario(@PathVariable int id) {
+	public ResponseEntity<Void> deletarFuncionario(@PathVariable BigInteger id) {
 		servico.deletarFuncionario(id);
 		return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
 	}
