+++ src/main/java/com/delegrego/exemplo_spring_boot_2/entity/DepartamentoEntity.java
@@ -1,5 +1,7 @@
 package com.delegrego.exemplo_spring_boot_2.entity;
 
+import java.math.BigInteger;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -29,7 +31,7 @@ public class DepartamentoEntity {
 	// IDENTITY é o tipo utilizado pelo MySQL
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "id_departamento")
-	private int idDepartamento;
+	private BigInteger idDepartamento;
 
 	// Restringe o atributo para não ser nulo e ter limite de 50 caracteres
 	@Column(name = "nm_departamento", nullable = false, length = 50)
