+++ src/main/java/com/delegrego/exemplo_spring_boot_2/repo/DepartamentoRepository.java
@@ -1,12 +1,13 @@
 package com.delegrego.exemplo_spring_boot_2.repo;
 
+import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.delegrego.exemplo_spring_boot_2.entity.DepartamentoEntity;
 
-public interface DepartamentoRepository extends JpaRepository<DepartamentoEntity, Integer> {
+public interface DepartamentoRepository extends JpaRepository<DepartamentoEntity, BigInteger> {
 
 	// Derived queries
